requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albumentations=1.4.0=pypi_0
antialiased-cnns=0.3=pypi_0
bcrypt=4.1.2=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
chainmap=1.0.3=pypi_0
chardet=3.0.4=pypi_0
charset-normalizer=3.3.2=pypi_0
clip=0.2.0=pypi_0
combomethod=1.0.12=pypi_0
contourpy=1.2.0=pypi_0
cryptography=42.0.5=pypi_0
cycler=0.12.1=pypi_0
et-xmlfile=1.1.0=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.49.0=pypi_0
fsspec=2024.2.0=pypi_0
googletrans=3.0.0=pypi_0
h11=0.9.0=pypi_0
h2=3.2.0=pypi_0
hpack=3.0.0=pypi_0
hstspreload=2024.3.1=pypi_0
httpcore=0.9.1=pypi_0
httpx=0.13.3=pypi_0
huggingface-hub=0.21.3=pypi_0
hyperframe=5.2.0=pypi_0
idna=2.10=pypi_0
imageio=2.34.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
nulltype=2.3.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=h7f8727e_0
options=1.4.10=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
paramiko=3.4.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
positional-encodings=6.0.3=pypi_0
pycparser=2.21=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-warmup=0.1.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rfc3986=1.5.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
scp=0.15.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.12.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
torch=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.38.2=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
